Moved to different package and removed unnecessary class.
