+++ server/src/main/java/de/uftos/repositories/solver/TimefoldSolver.java
@@ -1,4 +1,4 @@
-package de.uftos.solver;
+package de.uftos.repositories.solver;
 
 import ai.timefold.solver.core.api.solver.Solver;
 import ai.timefold.solver.core.api.solver.SolverFactory;
@@ -21,22 +21,22 @@ import de.uftos.dto.solver.TimetableProblemDto;
 import de.uftos.dto.solver.TimetableSolutionDto;
 import de.uftos.dto.ucdl.ConstraintDefinitionDto;
 import de.uftos.repositories.solver.SolverRepository;
-import de.uftos.solver.timefold.constraints.ConstraintDefinitionFactory;
-import de.uftos.solver.timefold.constraints.ConstraintDefinitionTimefoldInstance;
-import de.uftos.solver.timefold.constraints.ConstraintInstanceFactory;
-import de.uftos.solver.timefold.constraints.constraintinstances.ConstraintInstanceTimefoldInstance;
-import de.uftos.solver.timefold.domain.GradeTimefoldInstance;
-import de.uftos.solver.timefold.domain.LessonTimefoldInstance;
-import de.uftos.solver.timefold.domain.ResourceTimefoldInstance;
-import de.uftos.solver.timefold.domain.RoomTimefoldInstance;
-import de.uftos.solver.timefold.domain.StudentGroupTimefoldInstance;
-import de.uftos.solver.timefold.domain.StudentTimefoldInstance;
-import de.uftos.solver.timefold.domain.SubjectTimefoldInstance;
-import de.uftos.solver.timefold.domain.TagTimefoldInstance;
-import de.uftos.solver.timefold.domain.TeacherTimefoldInstance;
-import de.uftos.solver.timefold.domain.TimeslotTimefoldInstance;
-import de.uftos.solver.timefold.domain.TimetableSolutionTimefoldInstance;
-import de.uftos.solver.timefold.solver.ConstraintProviderTimefoldInstance;
+import de.uftos.repositories.solver.timefold.constraints.ConstraintDefinitionFactory;
+import de.uftos.repositories.solver.timefold.constraints.ConstraintDefinitionTimefoldInstance;
+import de.uftos.repositories.solver.timefold.constraints.ConstraintInstanceFactory;
+import de.uftos.repositories.solver.timefold.constraints.constraintinstances.ConstraintInstanceTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.GradeTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.LessonTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.ResourceTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.RoomTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.StudentGroupTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.StudentTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.SubjectTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TagTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TeacherTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TimeslotTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TimetableSolutionTimefoldInstance;
+import de.uftos.repositories.solver.timefold.solver.ConstraintProviderTimefoldInstance;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
