+++ server/src/main/java/de/uftos/repositories/solver/timefold/constraints/ConstraintDefinitionFactory.java
@@ -1,4 +1,4 @@
-package de.uftos.solver.timefold.constraints;
+package de.uftos.repositories.solver.timefold.constraints;
 
 import de.uftos.dto.ResourceType;
 import de.uftos.dto.solver.RewardPenalize;
@@ -11,18 +11,18 @@ import de.uftos.dto.ucdl.ast.OperatorDto;
 import de.uftos.dto.ucdl.ast.QuantifierDto;
 import de.uftos.dto.ucdl.ast.SetDto;
 import de.uftos.dto.ucdl.ast.ValueDto;
-import de.uftos.solver.timefold.domain.GradeTimefoldInstance;
-import de.uftos.solver.timefold.domain.LessonTimefoldInstance;
-import de.uftos.solver.timefold.domain.Number;
-import de.uftos.solver.timefold.domain.ResourceTimefoldInstance;
-import de.uftos.solver.timefold.domain.RoomTimefoldInstance;
-import de.uftos.solver.timefold.domain.StudentGroupTimefoldInstance;
-import de.uftos.solver.timefold.domain.StudentTimefoldInstance;
-import de.uftos.solver.timefold.domain.SubjectTimefoldInstance;
-import de.uftos.solver.timefold.domain.TagTimefoldInstance;
-import de.uftos.solver.timefold.domain.TeacherTimefoldInstance;
-import de.uftos.solver.timefold.domain.TimeslotTimefoldInstance;
-import de.uftos.solver.timefold.domain.TimetableSolutionTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.GradeTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.LessonTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.Number;
+import de.uftos.repositories.solver.timefold.domain.ResourceTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.RoomTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.StudentGroupTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.StudentTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.SubjectTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TagTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TeacherTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TimeslotTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.TimetableSolutionTimefoldInstance;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Iterator;
