+++ server/src/main/java/de/uftos/repositories/solver/timefold/constraints/ConstraintDefinitionTimefoldInstance.java
@@ -1,7 +1,7 @@
-package de.uftos.solver.timefold.constraints;
+package de.uftos.repositories.solver.timefold.constraints;
 
 import de.uftos.dto.solver.RewardPenalize;
-import de.uftos.solver.timefold.domain.ResourceTimefoldInstance;
+import de.uftos.repositories.solver.timefold.domain.ResourceTimefoldInstance;
 import java.util.List;
 import java.util.function.Function;
