+++ server/src/main/java/de/uftos/repositories/solver/timefold/domain/TeacherTimefoldInstance.java
@@ -1,4 +1,4 @@
-package de.uftos.solver.timefold.domain;
+package de.uftos.repositories.solver.timefold.domain;
 
 import ai.timefold.solver.core.api.domain.entity.PlanningEntity;
 import ai.timefold.solver.core.api.domain.lookup.PlanningId;
@@ -11,26 +11,27 @@ import java.util.List;
 import lombok.Getter;
 
 /**
- * This class models a room that can be used by the Timefold solver.
+ * This class models a teacher that can be used by the Timefold solver.
  */
 @PlanningEntity
 @Getter
-@JsonIdentityInfo(scope = RoomTimefoldInstance.class,
+@JsonIdentityInfo(scope = TeacherTimefoldInstance.class,
     generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
-public class RoomTimefoldInstance implements ResourceTimefoldInstance {
+public class TeacherTimefoldInstance implements ResourceTimefoldInstance {
 
   @PlanningId
   private final String id;
   private final List<TagTimefoldInstance> providedTagsList = new ArrayList<>();
-  @InverseRelationShadowVariable(sourceVariableName = "room")
+  private final List<SubjectTimefoldInstance> subjectList = new ArrayList<>();
+  @InverseRelationShadowVariable(sourceVariableName = "teacher")
   public List<LessonTimefoldInstance> lessonList;
 
-  public RoomTimefoldInstance(String id) {
+  public TeacherTimefoldInstance(String id) {
     this.id = id;
   }
 
   @Override
   public ResourceType getResourceType() {
-    return ResourceType.ROOM;
+    return ResourceType.TEACHER;
   }
 }
